Moved populate hashmap algorithm to utilities
